Move cmd/pivit/print.go -> pkg/pivit/print.go
